requirements.txt
atomicwrites==1.3.0
attrs==19.2.0
certifi==2019.9.11
chardet==3.0.4
coverage==4.5.4
cycler==0.10.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.6
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.10.3
factory-boy==2.12.0
Faker==2.0.2
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
kafka-python==1.4.7
kiwisolver==1.1.0
matplotlib==3.1.2
more-itertools==7.2.0
nltk==3.4.5
numpy==1.17.4
oauthlib==3.1.0
packaging==19.2
pandas==0.25.3
Pillow==6.2.1
pluggy==0.13.0
psycopg2==2.8.3
py==1.8.0
PyJWT==1.7.1
pyparsing==2.4.2
PySocks==1.7.1
pytest==5.2.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
text-unidecode==1.3
textblob==0.15.3
tweepy==3.8.0
urllib3==1.25.6
wcwidth==0.1.7
whitenoise==4.1.4
wordcloud==1.6.0
zipp==0.6.0
